requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
google-cloud-aiplatform[adk,agent_engines]==1.99.0
google-cloud-aiplatform==1.99.0
a2a-sdk==0.2.*
git+https://github.com/google/adk-python.git@main
streamlit
tabulate
platformdirs
Flask
gunicorn
scikit-learn
google-genai==1.21.1
pillow
pandas
google-cloud-bigquery[pandas]
google-cloud-bigquery-storage
google-cloud-firestore==2.21.*
altair==5.5.*
vl-convert-python
python-dotenv
cloudpickle==3.1.1
pydantic==2.11.*
yfinance
matplotlib
toolbox-core
uv
fastmcp==2.6.1